Refactor CustomZoomWindow to use integer percentages
Updated `CustomZoomWindowViewModel` to use `int?` for the `Number` property, ensuring zoom percentages are integers. Enhanced the `NumericUpDown` control in `CustomZoomWindow.axaml` with integer-specific properties (`Increment`, `ParsingNumberStyle`, `FormatString`) and adjusted its width for better UI. Simplified null-check logic in `OkButton_Click` in `CustomZoomWindow.axaml.cs` for clarity. Removed obsolete `NumberFormatter` comment and made minor formatting adjustments.

+++ Editor/ViewModels/CustomZoomWindowViewModel.cs
@@ -7,5 +7,5 @@ public partial class CustomZoomWindowViewModel : ObservableObject
     public int MaxPercentage { get; } = 9999;
 
     [ObservableProperty]
-    private double? _number = null;
+    private int? _number = null;
 }
